app/requirements.txt
# Base tools
setuptools>=68.0.0
wheel>=0.40.0

# Main dependencies
fastapi==0.109.2
uvicorn==0.27.1
motor==3.7.0
pydantic==2.10.3
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]~=1.7.4
python-multipart==0.0.9
python-dotenv==1.0.0
pydantic-settings==2.1.0